refactor(web): move entity renderers to shared diff helpers
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
